% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClusterCrit.R
\name{checkClusterCrit}
\alias{checkClusterCrit}
\title{Analyze the Effect of Correlation on Clustering Criteria}
\usage{
checkClusterCrit(spectra, corlvls = seq(0.9, 0.999, 0.002), clusters = NULL,
  altPlot = FALSE)
}
\arguments{
\item{spectra}{An object of class \code{Spectra}.
Should be shrinkwrapped, see the examples.}

\item{corlvls}{Numeric. A vector of values in (0...1).
The correlation matrix of the \code{Spectra} object will be filtered for each
of these values before further processing.}

\item{clusters}{Integer. An optional vector of values as long as
\code{corlvls}.  If not \code{NULL}, these values are
taken as the optimal number of clusters.}

\item{altPlot}{Logical. Should the alternate plot be made?
Only applies if \code{clusters == NULL}.
The default plot shows the criterion vs. the specified correlation values.
The alternative plot shows the size of the largest cluster vs. the 
number of clusters.
\code{altPlot} is forced to \code{TRUE} if only one correlation value
is given (since it is not possible to make the default plot in this case).}
}
\value{
A data frame with four columns:
\describe{
\item{cor}{the specified correlation levels}
\item{no.clust}{either the
optimal number of clusters found or the number specified by
\code{clusters}}
\item{maxclust}{the size of the largest cluster}
\item{crit}{the value of the criterion, which is maxclust/no.clust}
}
}
\description{
This function carries out an analysis designed to select the optimal
correlation value for use in a CLV analysis.  It is based roughly on Figure 5
of the reference.  The correlation matrix of the intensities in the
\code{Spectra} object are first filtered by the values in \code{corlvls}.
The resulting \code{Spectra} object is smaller as a result.  It is then
subjected to CLV analysis via \code{\link{clvSpectra}}.  In the process, the
size of the largest cluster is compared to the number of clusters, a criterion
recommended in the reference.  Please see Details.
}
\section{Details}{

There are two main ways to run this function.  (1) If \code{clusters} is \code{NULL},
and if \code{corlvls} has several values, each value
produces a delta plot with the optimal number of clusters marked (chosen by
\code{\link{findElbow}}, which may not be ideal for your data).  A final plot
shows the criterion vs. the correlation levels.  If \code{corlvls} has only one
value, \code{altPlot} is automatically set to \code{TRUE} and a plot of the
maximum cluster size vs. the number of clusters is made instead.  This approach
allows you to survey the results.  If you don't like the results or wish to
specify the number of clusters directly, then use the next approach.  (2)
If \code{clusters} is not \code{NULL}, then the specified number of clusters
is used (instead of looking for the optimal number) and a single plot is made,
showing the criterion vs. the correlation levels.  This is faster because only
one partition is computed for each value of \code{corlvls} instead of checking
a number of partitions (which defaults to 20 for each value of \code{corlvls}).

For the plots showing the criterion vs. the correlation levels, two vertical
dashed gray lines are show.  These mark the recommended minimum and maximum
values of the criterion, per the reference.  If they are the same value, a
message is sent to the console.
}

\section{Note}{

This function does a lot of heavy lifting and it is probably
best to run it in batch mode, especially for large data sets.
}
\examples{
\dontrun{
require("plyr")
require("ClustVarLV")
require("ChemoSpec")
data(metMUD2)

sw <- shrinkwrapSpectra(spectra = metMUD2, sigma = 0.02,
	thres = 4.5, xlim = c(1.25, 1.74), xlab = "ppm", plot = FALSE,
	main = expression(Shrinkwrap~of~metMUD2~sigma == 0.02))

cl <- seq(0.9, 0.99, 0.01) # 10 values, good for testing
tst <- checkClusterCrit(sw$Spectra, corlvls = cl)
tst2 <- checkClusterCrit(sw$Spectra, corlvls = cl, clusters = rep(4, length(cl)))
}
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
E. Vigneau and E. Qannari. "Clustering of variables around latent components."
\emph{Communications in Statistics - Simulation and Computation}, 32(4):1131-1150, 2003. 
For a good introduction, see E. Vigneau, M. Chen and E. Qannari
"ClustVarLV: An R Package for the Clustering of Variables Around Latent 
Variables" \emph{The R Journal} Vol. 7/2, December 2015
}
\keyword{multivariate}

