% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemoSpecMarkeR-package.R
\docType{package}
\name{ChemoSpecMarkeR-package}
\alias{ChemoSpecMarkeR}
\alias{ChemoSpecMarkeR-package}
\title{Functions to Identify Biomarkers in NMR Spectra}
\description{
A collection of functions for identifying biomarkers
in NMR spectra. Uses the \code{Spectra} class from the
ChemoSpec package.  There are two types of analyses possible with
this package.  1.  A strategy based on clustering of variables
around latent components (via the \code{ClustVarLV} package).
As part of this strategy, there are functions to "shrinkwrap" spectra
so that non-informative peaks are eliminated.
2. A strategy based on the STOCSY approach of Nicholson.  Please
refer to the functions for full references.
}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}

