% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFilterSpectra.R
\name{corFilterSpectra}
\alias{corFilterSpectra}
\title{Filter a Spectra Object Composed of NMR Data}
\usage{
corFilterSpectra(spectra, lvl = 0.9)
}
\arguments{
\item{spectra}{An object of class \code{Spectra}.
Should be shrinkwrapped, see the examples.}

\item{lvl}{Numeric. The correlation level.  One value in (0...1).}
}
\value{
An object of class \code{Spectra}.
}
\description{
This function computes the correlation matrix of
the intensities in the \code{Spectra} object.  Any 
pair of peaks with a correlation below the specified threshold
(argument \code{lvl}) 
are removed from the \code{Spectra} object, giving
a smaller object which is composed of only highly
positively correlated peaks.
}
\examples{

require("ClustVarLV")
require("ChemoSpec")
data(metMUD2)

sw <- shrinkwrapSpectra(spectra = metMUD2, sigma = 0.02,
thres = 4.5, xlim = c(1.25, 1.74), xlab = "ppm", plot = FALSE,
main = expression(Shrinkwrap~of~metMUD2~sigma == 0.02))

tst <- corFilterSpectra(sw$Spectra)
length(sw$Spectra$freq) # original no. of frequencies
length(tst$freq) # after filtering

}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
E. Vigneau and E. Qannari. "Clustering of variables around latent components."
\emph{Communications in Statistics - Simulation and Computation}, 32(4):1131-1150, 2003. 
For a good introduction, see E. Vigneau, M. Chen and E. Qannari
"ClustVarLV: An R Package for the Clustering of Variables Around Latent 
Variables" \emph{The R Journal} Vol. 7/2, December 2015
}
\keyword{multivariate}

