% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCLVclusters.R
\name{extractCLVclusters}
\alias{extractCLVclusters}
\title{Extract the Clusters Identified in a CLV Analysis of a Spectra Object Composed of NMR Data}
\usage{
extractCLVclusters(spectraORIG, spectraSW, clv = NULL, whichP = 1,
  plot = TRUE, ...)
}
\arguments{
\item{spectraORIG}{An object of class \code{Spectra}.  The data to be plotted.
Should not be shrink-wrapped.}

\item{spectraSW}{An object of class \code{Spectra}. The source of the frequencies
of interest.  Should be shrink-wrapped.  See \code{\link{shrinkwrapSpectra}}.}

\item{clv}{An object of class \code{"clv"}.  Created by
\code{\link{CLV}} or \code{\link{CLV_kmeans}} in package \code{ClustVarLV}.
In this case, created via \code{\link{clvSpectra}}.}

\item{whichP}{Integer.  The desired partition. See Details.}

\item{plot}{Logical.  Should a plot be made?}

\item{...}{Additional arguments to be passed to the plotting routines.
\code{xlim} would be a typical example, since you usually need to zoom
in to see what's going on.}
}
\value{
A list with elements equal to the number of clusters. Each element
contains the frequencies associated with that cluster.  Side effect is a plot.
}
\description{
This function extracts the clusters of frequencies identified in a CLV analysis
of a \code{Spectra} object containing NMR data.  The clusters of frequencies
are optionally presented as a dotplot for visual comparison.
Clusters of frequences (peaks) may be
related by being part of the same spin system/compound,
or depending upon the experimental design, the peaks
identified may be related because they arise from
separate compounds which co-vary because they are on
related metabolic pathways.
}
\section{Details}{

Objects of class \code{"clv"} contain different
entries if they were created by function \code{CLV} or \code{CLV_kmeans}.
For \code{mode = "clv"}
a number of partitions are returned, while for 
\code{mode = "kmeans"} only a single partition is returned.
In the latter case argument
\code{whichP} is ignored.
}
\examples{

require("ClustVarLV")
require("ChemoSpec")
data(metMUD2)

# Shrinkwrap the spectra
sw <- shrinkwrapSpectra(spectra = metMUD2, sigma = 0.01,
thres = 4.85, xlim = c(1.25, 1.74), plot = FALSE)

# Now CLV analysis
clv1 <- clvSpectra(sw$Spectra)
plot(clv1, type = "delta") # suggests 4 clusters
freqs <- extractCLVclusters(metMUD2, sw$Spectra, clv = clv1, whichP = 4,
main = "Peak Clusters in metMUD2")

}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
E. Vigneau and E. Qannari. "Clustering of variables around latent components."
\emph{Communications in Statistics - Simulation and Computation}, 32(4):1131-1150, 2003. 
For a really good introduction, see E. Vigneau, M. Chen and E. Qannari
"ClustVarLV: An R Package for the Clustering of Variables Around Latent 
Variables" \emph{The R Journal} Vol. 7/2, December 2015
}
\seealso{
\code{\link{plotSpectraCLV}}
}
\keyword{multivariate}

