% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptimumPartition.R
\name{getOptimumPartition}
\alias{getOptimumPartition}
\title{Find the Optimum Number of Partitions in a CLV Analysis}
\usage{
getOptimumPartition(clv, plot = TRUE)
}
\arguments{
\item{clv}{An object of class \code{"clv"}.}

\item{plot}{Logical. Should a plot be made?
The plot is identical to that produced by
\code{plot(clv_object, type = "delta")}
except that a red diamond marks the
optimal partition.}
}
\value{
Integer.  The optimal partition.
}
\description{
This function provides an automated means of selecting the
optimum number of clusters in a clv analysis, which is
useful in certain applications.  The procedure is to
identify the elbow in the delta plot derived from a 
call to \code{\link{CLV}}.  The elbow is the delta value
which is farthest (orthogonally) from a line connecting
the delta values for the largest number of partitions to the
one for the smallest number of partitions.
}
\section{Warning}{

As an automated method, this function returns a consistent answer.
However, that answer may not be ideal for any given data set.  It
is advisable to inspect the delta plots when using this function
to ensure that the concept here works well for your data sets.
}
\examples{

require("ClustVarLV")
data("AUPA_psycho")
res1<- CLV(AUPA_psycho, method = "directional", sX = TRUE)
plot(res1, type = "delta")
part <- getOptimumPartition(res1)

data("apples_sh")
res2 <- CLV(apples_sh$pref, method = "local")
plot(res2, type = "delta")
part <- getOptimumPartition(res2)

require("FactoMineR")
data(wine)
X.quanti <- wine[, 3:29]
X.quali <- wine[, 1:2]
Xbig <- cbind(scale(X.quanti), stand_quali((X.quali)))
res3 <- CLV(Xbig, method = "directional", sX = FALSE)
plot(res3, type = "delta")
part <- getOptimumPartition(res3)

}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
E. Vigneau and E. Qannari. "Clustering of variables around latent components."
\emph{Communications in Statistics - Simulation and Computation}, 32(4):1131-1150, 2003. 
For a good introduction, see E. Vigneau, M. Chen and E. Qannari
"ClustVarLV: An R Package for the Clustering of Variables Around Latent 
Variables" \emph{The R Journal} Vol. 7/2, December 2015
}
\seealso{
\code{\link{findElbow}} for more about how the process works.
}

