% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkwrapSpectra.R
\name{shrinkwrapSpectra}
\alias{shrinkwrapSpectra}
\title{Shrinkwrap a Spectra Object Composed of NMR Data}
\usage{
shrinkwrapSpectra(spectra, thres = 2, plot = TRUE, method = "theoretical",
  sigma = 5e-04, W = 2000, ...)
}
\arguments{
\item{spectra}{An object of class \code{Spectra}.}

\item{thres}{Numeric.  Any peaks with a maximum below this value
will be discarded.  Determine by inspection (see examples).}

\item{plot}{Logical.  Option for plot showing bucketed peaks
on the original spectrum.}

\item{method}{One of \code{c("theoretical", "empirical")}, giving the method for
computing the second derivative.}

\item{sigma}{Only applies to \code{method = "theoretical"}.  Numeric.
Peak width at half maximum of the curve (FWHM).  Equivalent to
2 * gamma, the half-width at half-maximum (HWHM).
In simpler terms, \code{sigma} should correspond to typical peak widths in
the spectrum. The units of \code{sigma}
are the units of the frequency values in the \code{Spectra Object}.
The default value is 0.0005 ppm, as recommended in the reference.
You probably need to adjust this parameter.  See Details and the examples.}

\item{W}{Only applies to \code{method = "theoretical"}.  Integer.
The number of data points in the window used to
compute the second derivative.  Due to the shape of a Lorentzian
curve, the second derivative declines dramatically beyond a
certain point and time can be saved by not computing values
outside a certain range.  The reference recommends a value of 2000
for a typical NMR data set.}

\item{...}{Additional arguments to be passed to the plotting routines.
Most likely you will want to adjust \code{xlim} so you can see 
details in the plots.}
}
\value{
A list consisting of a modified \code{Spectra} object, and
the indices of the peak boundaries.  Indices correspond to the
original \code{Spectra} object, not the one returned.
}
\description{
This function identifies the central part of an NMR peak using
a modification of the convolution product / second
derivative method described in the
reference.  The rest of the frequencies are then discarded,
resulting in a much smaller data set.
To ensure that all peaks in a data set are
represented, the spectra are added together as a first step.
In this context, a peak means any peak, ignoring any larger structure.
For instance, all three
peaks in a triplet will be identified as separate peaks.  This
facilitates database searching in a subsequent step (and
in a crowded spectrum there is no knowing what goes with what
anyway).
}
\section{Details}{

The theoretical method is based upon
equation 2 in the reference, and requires an assumption about the value
of sigma.  However, this function does not extend the zero crossings
as described in the reference because the same effect can be accomplished
by adjusting the value of \code{sigma}.  The results of the second
derivative calculation are very sensitive to the value of \code{sigma}.
It is recommended that you experiment to find a suitable value.
You can do so using this function and examining a few regions
of interest to get \code{thres} and \code{sigma} the way you want them.
The empirical method uses the
Savitzky-Golay second derivative and makes no assumptions, so 
it is more robust.  This method also typically identifies many
more peaks than the theoretical method.
}
\examples{

require("ChemoSpec")
data(metMUD2)

# Inspect to determine suitable threshold
s <- colSums(metMUD2$data)
x <- metMUD2$freq
plot(x, s, type = "l", ylim = c(4, 8),
	xlab = "ppm", ylab = "colSums", main = "Checking the Threshold")
th <- 4.5
abline(h = th, col = "green")

# Shrinkwrap using theoretical method
# Note the effect of sigma on the results

sw <- shrinkwrapSpectra(spectra = metMUD2, sigma = 0.01,
thres = th, xlim = c(1.25, 1.74), xlab = "ppm",
main = expression(Shrinkwrap~of~metMUD2~sigma == 0.01))

sw <- shrinkwrapSpectra(spectra = metMUD2, sigma = 0.02,
thres = th, xlim = c(1.25, 1.74), xlab = "ppm",
main = expression(Shrinkwrap~of~metMUD2~sigma == 0.02))

sw <- shrinkwrapSpectra(spectra = metMUD2, sigma = 0.01,
thres = th, xlim = c(0.85, 1.1), xlab = "ppm",
main = expression(Shrinkwrap~of~metMUD2~sigma == 0.01))

sw <- shrinkwrapSpectra(spectra = metMUD2, sigma = 0.02,
thres = th, xlim = c(0.85, 1.1), xlab = "ppm",
main = expression(Shrinkwrap~of~metMUD2~sigma == 0.02))

# Compare to empirical method (finds more peaks at a given thres)

sw <- shrinkwrapSpectra(spectra = metMUD2, thres = 5,
xlim = c(0.85, 1.1), method = "empirical", xlab = "ppm",
main = "Shrinkwrap of metMUD2, method = empirical")

\dontrun{

# A much larger data set
data(SrE.NMR)

# Inspect and set the threshold
s <- colSums(SrE.NMR$data)
x <- SrE.NMR$freq
plot(x, s, type = "l", ylim = c(-1e6, 1e7),
	xlab = "ppm", ylab = "colSums", main = "Checking the Threshold")
th <- 5e5
abline(h = th, col = "green")

sw <- shrinkwrapSpectra(spectra = SrE.NMR, sigma = 0.01,
thres = th, xlim = c(1.2, 1.4), xlab = "ppm",
main = expression(Shrinkwrap~of~SrE.NMR~sigma == 0.01))

sw <- shrinkwrapSpectra(spectra = SrE.NMR, sigma = 0.02,
thres = th, xlim = c(1.2, 1.4), xlab = "ppm",
main = expression(Shrinkwrap~of~SrE.NMR~sigma == 0.02))

sw <- shrinkwrapSpectra(spectra = SrE.NMR,
thres = th, xlim = c(1.2, 1.4), xlab = "ppm", method = "empirical",
main = "Shrinkwrap of SrE.NMR, method = empirical")
}
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
Daniel Jacob, Catherine Deborde, and Annick Moing.
"An Efficient Spectra Processing Method for Metabolite Identification
from 1H-NMR Metabolomics Data."
Analytical and Bioanalytical Chemistry vol. 405 (2013) pgs. 5049-5061.
DOI: 10.1007/s00216-013-6852-y
}
\keyword{utilities}

